Split out isPostPublished
